Clarify distance between NaNs, INFINITY
Distance to a NaN -- even from another NaN -- now always returns
the maximum distance.

Distance to INFINITY returns the maximum distance unless both values
are infinite and also have matching signs.

If this is not the desired behavior in our use-case, we can build a
wrapper function to pre-check for these corner-cases.